refactor(google-genai): split config client and tools
